[Update] TableViewCell 디자인 수정, OrderButtonsView 텍스트 색상변경, 임시로 기존의 CartView를 ModalTestView로 이동
