+++ nbc-popcon/nbc-popcon/Sources/Cells/CartItemButtonCell.swift
@@ -22,22 +22,25 @@ class CartItemButtonCell: UITableViewCell {
     private let kiosk = Kiosk()
     private var itemQuantity = 1.0
     
+    
+    //MARK: -
     // 상품 imageView
     private let cartItemImageView: UIImageView = {
         let imageView = UIImageView()
-        imageView.image = UIImage(systemName: "wifi")
+        imageView.image = UIImage(systemName: "trash")
         imageView.contentMode = .scaleAspectFit
         imageView.tintColor = ThemeColors.black
+        //imageView.backgroundColor = .gray
         return imageView
     }()
-    
+    //MARK: -
     // 상품이름 Label
     private let cartItemLabel: UILabel = {
         let label = UILabel()
         label.text = "symbol name"
-        label.font = ThemeFonts.h3
+        label.font = ThemeFonts.h2
         label.textColor = ThemeColors.black
-        label.textAlignment = .center
+        label.textAlignment = .right
         return label
     }()
     
@@ -45,42 +48,11 @@ class CartItemButtonCell: UITableViewCell {
     private let cartItemPriceLabel: UILabel = {
         let label = UILabel()
         label.text = "10,000 원"
-        label.font = ThemeFonts.h3
-        label.textColor = ThemeColors.blue
-        label.textAlignment = .center
-        return label
-    }()
-    
-    //  상품 갯수 Label
-    lazy var cartItemQuantityLabel: UILabel = {
-        let label = UILabel()
-        label.text = "x \(Int(itemQuantity))"
-        label.font = ThemeFonts.h1
-        label.textColor = ThemeColors.black
-        label.textAlignment = .center
-        return label
-    }()
-    
-    // stapper
-    lazy var stepper: UIStepper = {
-        let stepper = UIStepper()
-        stepper.value = itemQuantity
-        stepper.stepValue = 1.0
-        stepper.minimumValue = 0.0
-        stepper.maximumValue = 50.0
-        return stepper
-    }()
-    
-    //  상품 합계 가격 Label
-    lazy var cartItemTotalPriceLabel: UILabel = {
-        let label = UILabel()
-        label.text = "총 40,000 원"
         label.font = ThemeFonts.h2
-        label.textColor = ThemeColors.black
+        label.textColor = ThemeColors.blue
         label.textAlignment = .right
         return label
     }()
-    
     // 전체 삭제 Button
     private let totalDeleteButton: UIButton = {
         let button = UIButton()
@@ -93,38 +65,73 @@ class CartItemButtonCell: UITableViewCell {
         button.contentHorizontalAlignment = .right
         return button
     }()
+    //MARK: -
+    //  상품 갯수 Label
+    lazy var cartItemQuantityLabel: UILabel = {
+        let label = UILabel()
+        label.text = "주문수량 : \(Int(itemQuantity)) 개"
+        label.font = ThemeFonts.h3
+        label.textColor = ThemeColors.black
+        label.textAlignment = .center
+        return label
+    }()
     
-    // 상품 이미지, 이름, 단가 verticalStackView
+    // stapper
+    lazy var stepper: UIStepper = {
+        let stepper = UIStepper()
+        stepper.value = itemQuantity
+        stepper.stepValue = 1.0
+        stepper.minimumValue = 0.0
+        stepper.maximumValue = 50.0
+        stepper.backgroundColor = .orange
+        return stepper
+    }()
+    
+    //  상품 합계 가격 Label
+    lazy var cartItemTotalPriceLabel: UILabel = {
+        let label = UILabel()
+        label.text = "= 20,000 원"
+        label.font = ThemeFonts.h3bold
+        label.textColor = ThemeColors.black
+        label.textAlignment = .center
+        return label
+    }()
+    //MARK: -
+    // 이름, 단가 삭제버튼 verticalStackView
     private let itemVerticalstackView: UIStackView = {
         let stackView = UIStackView()
         stackView.axis = .vertical
         stackView.spacing = 5
         stackView.distribution = .fill
-        stackView.alignment = .center
+        stackView.alignment = .trailing
+        //stackView.backgroundColor = .red
         return stackView
     }()
     
-    // 갯수와 stepper horizontalView
-    private let countHorizontalView: UIView = {
+    // image와 오른쪽 View
+    private let itemHorizonView: UIView = {
         let view = UIView()
+        //view.backgroundColor = .green
         return view
     }()
     
-    // 카운트, 합계금액, 삭제버튼 verticalStackView
-    private let calculateVerticalstackView: UIStackView = {
+    // 갯수와 stepper 합계금액 horizontalView
+    private let calculateHorizontalView: UIStackView = {
         let stackView = UIStackView()
-        stackView.axis = .vertical
-        stackView.spacing = 10
+        stackView.axis = .horizontal
+        stackView.spacing = 5
         stackView.distribution = .fill
+        stackView.backgroundColor = ThemeColors.grey6
         return stackView
     }()
     
-    // 전체 horizontalView
-    private let horizontalView: UIView = {
+    // 상품 imageView
+    private let cellView: UIView = {
         let view = UIView()
+        //view.backgroundColor = .cyan
         return view
     }()
- 
+    
     //MARK: - setting
     
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
@@ -142,91 +149,96 @@ class CartItemButtonCell: UITableViewCell {
     // TableViewCell layout 설정
     private func setupTableViewCellLayout() {
         
-        itemVerticalstackView.addArrangedSubview(cartItemImageView)
+        itemHorizonView.addSubview(cartItemImageView)
         itemVerticalstackView.addArrangedSubview(cartItemLabel)
         itemVerticalstackView.addArrangedSubview(cartItemPriceLabel)
-        countHorizontalView.addSubview(cartItemQuantityLabel)
-        countHorizontalView.addSubview(stepper)
-        calculateVerticalstackView.addArrangedSubview(countHorizontalView)
-        calculateVerticalstackView.addArrangedSubview(cartItemTotalPriceLabel)
-        calculateVerticalstackView.addArrangedSubview(totalDeleteButton)
-        horizontalView.addSubview(itemVerticalstackView)
-        horizontalView.addSubview(calculateVerticalstackView)
-
-        contentView.addSubview(horizontalView)
-      
-        // 전체를 감싼 StackView
-        horizontalView.snp.makeConstraints {
+        itemVerticalstackView.addArrangedSubview(totalDeleteButton)
+        itemHorizonView.addSubview(itemVerticalstackView)
+        calculateHorizontalView.addArrangedSubview(cartItemQuantityLabel)
+        calculateHorizontalView.addArrangedSubview(stepper)
+        calculateHorizontalView.addArrangedSubview(cartItemTotalPriceLabel)
+        contentView.addSubview(cellView)
+        contentView.addSubview(itemHorizonView)
+        contentView.addSubview(calculateHorizontalView)
+        
+        // 전체를 감싼 cellView
+        cellView.snp.makeConstraints {
             $0.top.equalTo(contentView.snp.top)
             $0.bottom.equalTo(contentView.snp.bottom)
             $0.leading.equalTo(contentView.snp.leading)
             $0.trailing.equalTo(contentView.snp.trailing)
         }
         
-        // 왼쪽의 이미지, 이름, 단가 verticalStackView
-        itemVerticalstackView.snp.remakeConstraints {
-            $0.leading.equalToSuperview().inset(ThemeNumbers.padding)
-            $0.top.equalToSuperview().inset(ThemeNumbers.padding)
-            $0.bottom.equalToSuperview().inset(ThemeNumbers.padding)
-            $0.width.equalTo(130)
-        }
-            
-        // 오른쪽 수량과 스태퍼 horizontalView
-        countHorizontalView.snp.makeConstraints {
-            $0.trailing.equalToSuperview().inset(10)
-            $0.top.equalToSuperview()
+        itemHorizonView.snp.makeConstraints {
+            $0.top.leading.trailing.equalToSuperview().inset(10)
+            $0.bottom.equalTo(calculateHorizontalView.snp.top)
         }
         
-        // 수량, 가격, 삭제버튼 verticalStackView
-        calculateVerticalstackView.snp.makeConstraints {
-            $0.top.equalToSuperview().inset(40)
-            $0.trailing.equalToSuperview().inset(ThemeNumbers.padding)
-            $0.bottom.equalToSuperview().inset(ThemeNumbers.padding)
-            $0.width.equalTo(150)
-        }
-         
         // 아이템 이미지 Layout
         cartItemImageView.snp.makeConstraints {
-            $0.height.equalTo(100)
+            $0.top.bottom.equalToSuperview().inset(5)
+            $0.leading.equalToSuperview().inset(20)
             $0.width.equalTo(100)
+        }
+        
+        // 오른쪽 이름, 단가 삭제버튼 verticalStackView
+        itemVerticalstackView.snp.makeConstraints {
+            $0.trailing.equalToSuperview().inset(10)
+            $0.top.equalToSuperview().inset(10)
+            $0.bottom.equalTo(calculateHorizontalView.snp.top)
+            $0.width.equalTo(150)
+        }
+        
+        // 아래 수량과 스태퍼 합계 horizontalView
+        calculateHorizontalView.snp.makeConstraints {
+            $0.leading.trailing.equalToSuperview()
+            $0.bottom.equalToSuperview()
+            $0.height.equalTo(50)
         }
         
         // 아이템 이름 Layout
         cartItemLabel.snp.makeConstraints {
             $0.height.equalTo(ThemeNumbers.padding)
             $0.width.equalTo(itemVerticalstackView.snp.width)
-            $0.leading.equalTo(itemVerticalstackView.snp.leading).inset(10)
+            $0.trailing.equalTo(itemVerticalstackView.snp.trailing)
         }
         
         // 아이템 단가 Layout
         cartItemPriceLabel.snp.makeConstraints {
-            $0.height.equalTo(ThemeNumbers.padding + 5)
-            $0.leading.equalTo(itemVerticalstackView.snp.leading).inset(10)
-        }
-        
-        // 아이템 선택된 갯수 Layout
-        cartItemQuantityLabel.snp.makeConstraints {
-            $0.leading.equalToSuperview()
-            $0.centerY.equalToSuperview()
-            $0.height.equalToSuperview()
-            $0.width.equalTo(55)
-        }
-        
-        // stepper Layout
-        stepper.snp.makeConstraints {
-            $0.trailing.equalTo(countHorizontalView.snp.trailing)
-            $0.centerY.equalToSuperview()
-        }
-        
-        // 선택된 가격 Layout
-        cartItemTotalPriceLabel.snp.makeConstraints {
-            $0.height.equalTo(28)
+            $0.height.equalTo(ThemeNumbers.padding)
+            $0.trailing.equalTo(itemVerticalstackView.snp.trailing)
         }
         
         // 전체 삭제 버튼 Layout
         totalDeleteButton.snp.makeConstraints {
             $0.height.equalTo(22)
         }
+        
+        // 아이템 선택된 갯수 Layout
+        cartItemQuantityLabel.snp.makeConstraints {
+            $0.leading.equalToSuperview().inset(10)
+            $0.centerY.equalToSuperview()
+            $0.height.equalToSuperview()
+            $0.width.equalTo(140)
+        }
+        
+        // stepper Layout
+        stepper.snp.makeConstraints {
+            $0.centerY.equalToSuperview()
+              $0.leading.equalTo(cartItemQuantityLabel.snp.trailing).offset(10)
+              $0.width.equalTo(94)
+              $0.height.equalTo(30)
+        }
+        
+        // 선택된 아이템 가격 Layout
+        cartItemTotalPriceLabel.snp.makeConstraints {
+            $0.centerY.equalToSuperview()
+            $0.height.equalToSuperview()
+            $0.trailing.equalToSuperview()
+            $0.width.equalTo(140)
+        }
+        
+        
     }
 }
 
@@ -244,10 +256,10 @@ extension CartItemButtonCell {
         let newValue = sender.value
         if newValue > itemQuantity {
             itemQuantity = newValue
-            self.cartItemQuantityLabel.text = "x \(Int(self.itemQuantity))"
+            self.cartItemQuantityLabel.text = "주문수량 : \(Int(itemQuantity)) 개"
         } else if newValue < itemQuantity {
             itemQuantity = newValue
-            self.cartItemQuantityLabel.text = "x \(Int(self.itemQuantity))"
+            self.cartItemQuantityLabel.text = "주문수량 : \(Int(itemQuantity)) 개"
         }
     }
 }
